+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+### Added
+
+- Support `image/avif` media type.
+
 ### Changed
 
 - Accept forwarded `Update(Note)` activities.